+++ NEWS.md
@@ -1,6 +1,6 @@
 # reuseme (development version)
 
-* `screenshot()` has better support for Quarto blogs.
+* `screenshot()` has better support for Quarto blogs, and gains a `proj` argument.
 
   - The image will be saved in the active `posts/` folder, queried using `rstudioapi::documentPath()`.